requirements.txt
beautifulsoup4
langchain
langgraph
langchain_community
langchain_core
python-dotenv
streamlit
langchain_postgres
google-genai
langchain-google-genai
supabase
openai
rich
psycopg2-binary